Fix table indexing. Note about IMU
The IMUs seem to cut out sometimes. As long as I don't touch them, I can run them for a really long time, but once I start moving them around, this becomes less reliable and can freeze up the whole system due to some problems in the HAL I2C driver (not handled by silicon bug fix).
I've concluded that these issues only arise when the wires lose a good connection, i.e. power, SDA, and/or SCL. Any of these can cause a freeze if they are tampered with to the extent that a bit may be lost.
Should really add a watchdog to handle these cases. Would need to figure out what to do about the camera LED if it is frequently resetting though...