Add WDT
Plan: use watchdog that generates an interrupt, and if it's our first time there, reset the I2C peripherals if they are in an erroneous state (busy, etc). Otherwise, reset.

APB1: 42 MHz

tWWDG = (1/42000) * 4096 * (2^WDGTB[1:0]) * (T[5:0] + 1)

T[6:0] are a downcounter. Reset/EWI occurs when T[6] becomes 0

If we choose WDGTB[1:0] = 1 and T[5:0] = 40, we get:
tWWDG = (1/42000) * 4096 * (2^2) * (32 + 1) = 12.873 ms