Interpolate NaNs in array when doing analyses
If we do not replace the NaNs, then they will mess up the flow (e.g. plots won't be made past the time of the first NaN).

Right now, linear interpolation is used to find reasonable guesses for the missing data. Seems to work fine, but further investigation should be undertaken.