Increase timer task priority so that status LED and flag are prioritized over sensor I/O
This was always the intention, however, I misunderstood Cube's priority interface. It turns out that if you have max priorities = 7, then 6 maps to osPriorityRealtime, 5 maps to osPriorityHigh, 4 maps to osPriorityAboveNormal, etc.

It's just weird that they make you choose task priorities based on the *name* of the priority, while they make you choose the software timer task priority based on a numbering scheme that doesn't correspond to the enum values used in the CMSIS API.