Start working on calibration functions
Not sure why, but the multiplicative factors don't seem to be needed. My only explanation is that the z (and possibly y?) axis are pointing opposite of the way I was expecting. Based on the pictures we have of the IMU wiring in /docs, I thought the z-axis would have been pointing down for both sensors. But either that's not the case, or I'm missing something...