Fix issues with calibration flow. Fix bool argparse bug. IMPORTANT note in description
So basically it turns out that all the IMU acceleration values are off by a factor of -1. This means that there needs to be a change in the embedded code that multiplies ax, ay, az by -1. This will make the acceleration values received by the computer log consistent with the coordinate system described in the sensor datasheet.

Once this change is made in embedded, the multiplication setting in settings.ini needs to be changed to:
mult_z = -1
mult_y = -1
mult_x = 1

Which is the negative of what it is right now.

This will not solve any functional issues per se, but it's a good improvement for consistency