Use only one IMU thread
Seems to have fixed all the issue where the IMU would sometimes report garbage values.

Since the last change, the issue noticed was that sometimes the base gyro readings would all be -0.01 (rounded). When I looked at the full decimal expansion, it turned out to be -0.00762962736. This translates to receiving 0xFFFF from the IMU when reading each of the gyros, which is definitely an error. For some reason, putting the IMUs in the same thread fixes this issue. I'm still investigating the root cause for this issue but want to make sure I commit the progress on the fix since it is promising