Add ability to enable and disable servo control and IMU sensing. Tested.
Currently these commands are only being used during playback and recording to ENABLE required modules. For example, when we enter playback mode on the PC script, it first sends a command to the microcontroller to enable servos. Similarly, when we enter recording mode on the PC script, it first sends a command to enable IMU sensing. This is done so that the user of the PC script doesn't have to issue more than one command to initiate a test in a desired mode. Hence, although the servo and IMU enable/disable commands are available directly from the command-line, their effects will always be overwritten as soon as recording or playback command is issued.

In the future, we could also send a command to disable servos when we are recording, and send a command to disable IMU sensing when we are in playback mode. Commands could also be added to enable/disable transmissions from the MCU to the PC, since these transmissions are discarded when in playback mode and are only needed in recording mode.

At the moment, the microcontroller disables servo actuation when it is first powered up, so it basically defaults into recording mode, and it requires a command to start servo actuation. This is a safety feature more than anything else, because it means that upon powering up the MCU, the servos will not automatically start moving